fix: retirado tipo double

+++ src/main/java/com/cc/vendas/aplicacao/dto/entrada/AtualizarVeiculoInput.java
@@ -1,10 +1,12 @@
 package com.cc.vendas.aplicacao.dto.entrada;
 
+import java.math.BigDecimal;
+
 public record AtualizarVeiculoInput(
         String marca,
         String modelo,
         String cor,
         Integer ano,
-        Double preco
+        BigDecimal preco
 ) {
 }
