+++ src/main/java/com/cc/vendas/aplicacao/dto/entrada/ConfirmacaoPagamentoInput.java
@@ -2,11 +2,12 @@ package com.cc.vendas.aplicacao.dto.entrada;
 
 import com.cc.vendas.shared.StatusPagamento;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record ConfirmacaoPagamentoInput(
         UUID vendaId,
-        Double valor,
+        BigDecimal valor,
         String codigoPagamento,
         StatusPagamento statusPagamento
 ) {
